+++ src/index.d.ts
@@ -1,8 +1,7 @@
 declare type FnT<T> = {
     (value: T): T;
 };
-declare function typedResolver<T>(initValue: T, callbacks: FnT<T>[]): T;
-declare function resolver(initValue: any, callbacks: FnT<any>[]): any;
-export default resolver;
-export { resolver, typedResolver };
+declare function series<T>(initValue: T, callbacks: FnT<T>[]): T;
+declare function parallel<T>(initValue: T, fns: FnT<T>[]): T[];
+export { parallel, series };
 //# sourceMappingURL=index.d.ts.map