+++ src/index.spec.ts
@@ -1,49 +1,16 @@
-import {resolver, typedResolver} from "./index";
+import {parallel, series} from "./index";
 
-describe('callback-chain-resolver default resolver', () => {
-    const incrementBy1 = (x: number) => x + 1;
-    const decrementBy1 = (x: number) => x - 1;
-    const double = (x: number) => 2 * x;
-    const square = (x: number) => x * x;
-    const cube = (x: number) => x * square(x);
+const incrementBy1 = (x: number) => x + 1;
+const decrementBy1 = (x: number) => x - 1;
+const double = (x: number) => 2 * x;
+const square = (x: number) => x * x;
+const cube = (x: number) => x * square(x);
+const toString = (x: number) => x.toString();
 
-    it('resolver() resolves successfully', () => {
-        [
-            {
-                callbackArray: [incrementBy1, incrementBy1, incrementBy1, double, square],
-                initValue: 1,
-                expected: 64
-            },
-            {
-                callbackArray: [double, square, decrementBy1, decrementBy1, decrementBy1],
-                initValue: 1,
-                expected: 1
-            },
-            {
-                callbackArray: [incrementBy1, cube, decrementBy1],
-                initValue: 1,
-                expected: 7
-            },
-            {
-                callbackArray: [decrementBy1, incrementBy1],
-                initValue: 1,
-                expected: 1
-            }
-        ].forEach(testCase => {
-            expect(resolver(testCase.initValue, testCase.callbackArray)).toEqual(testCase.expected);
-        });
-    });
-});
+describe('series', () => {
 
-describe('callback-chain-resolver typedResolver', () => {
-    const incrementBy1 = (x: number) => x + 1;
-    const decrementBy1 = (x: number) => x - 1;
-    const double = (x: number) => 2 * x;
-    const square = (x: number) => x * x;
-    const cube = (x: number) => x * square(x);
-    // const toString = (x: number) => x.toString();
 
-    it('typedResolver() resolves successfully', () => {
+    it('series<number>() resolves successfully', () => {
         [
             {
                 callbackArray: [incrementBy1, incrementBy1, incrementBy1, double, square],
@@ -66,7 +33,60 @@ describe('callback-chain-resolver typedResolver', () => {
                 expected: 1
             }
         ].forEach(testCase => {
-            expect(typedResolver(testCase.initValue, testCase.callbackArray)).toEqual(testCase.expected);
+            expect(series<number>(testCase.initValue, testCase.callbackArray)).toEqual(testCase.expected);
         });
+    });
+
+    // won't run
+    // it('series<string>() fails for number returning functions', () => {
+    //     // TS2345: Argument of type 'number' is not assignable to parameter of type 'string'.
+    //
+    //     [
+    //         {
+    //         callbackArray: [incrementBy1, incrementBy1, incrementBy1, double, square],
+    //         initValue: 1,
+    //         expected: 64
+    //     }
+    //     ].forEach(t => {
+    //         expect(series<string>(t.initValue, t.callbackArray)).toEqual(t.expected);
+    //     })
+    // });
+
+    // will run
+    it('series<string>() fails for number returning functions', () => {
+        // TS2345: Argument of type 'number' is not assignable to parameter of type 'string'.
+
+        [
+            {
+                callbackArray: [incrementBy1, incrementBy1, incrementBy1, double, square, toString],
+                initValue: 1,
+                expected: '64'
+            }
+        ].forEach(t => {
+            expect(series<any>(t.initValue, t.callbackArray)).toEqual(t.expected);
+        });
+    });
+});
+
+describe('parallel', () => {
+    it('maps functions to input values', () => {
+        const testCase = {
+            callbackArray: [incrementBy1, incrementBy1, incrementBy1, double, square],
+            initValue: 1,
+            expected: [2, 2, 2, 2, 1]
+        };
+        const actual = parallel<number>(testCase.initValue, testCase.callbackArray);
+        const expected = testCase.expected;
+        expect(actual).toEqual(expected);
+    });
+    it('maps mixed functions to input values', () => {
+        const testCase = {
+            callbackArray: [incrementBy1, toString, decrementBy1, double, square, (x: number) => series(x, [cube, decrementBy1, cube])],
+            initValue: 2,
+            expected: [3, '2', 1, 4, 4, 343]
+        };
+        const actual = parallel<any>(testCase.initValue, testCase.callbackArray);
+        const expected = testCase.expected;
+        expect(actual).toEqual(expected);
     });
 });